requirements.txt
# Ingestão
pandas==2.2.2
duckdb==0.10.3

# dbt
dbt-core==1.7.9
dbt-duckdb==1.7.2

# Orquestração
apache-airflow==2.9.1

# Utilitários
python-dotenv==1.0.1
loguru==0.7.2